Add ServerBase for shared Node and Ruby Sass code Add cleanup routine for Sass Server Update Sass to 3.4.7
